refactor: Move match methods to ItemFilterService & make ItemSearchQuery a record
