+++ common/src/main/resources/assets/wynntils/lang/en_us.json
@@ -482,7 +482,7 @@
   "feature.wynntils.itemFavorite.lootChestCloseOverride.description": "After how many ESC presses should the chest be closed? Set to 0 to disable.",
   "feature.wynntils.itemFavorite.lootChestCloseOverride.name": "Loot Chest Close Override",
   "feature.wynntils.itemFavorite.name": "Favorite Items",
-  "feature.wynntils.itemFilter.level.invalidRange": "%s: Should be an integer or a range, seperated by a dash",
+  "feature.wynntils.itemFilter.level.invalidRange": "%s: Should be an integer or a range, separated by a dash",
   "feature.wynntils.itemFilter.level.maxLessThanMin": "%s: Min must be less than max",
   "feature.wynntils.itemFilter.level.usage": "<level> or <min>-<max>",
   "feature.wynntils.itemFilter.profession.invalidProfession": "%s: invalid profession",
